refactor: organize data types
