+++ src/schema/index.ts
@@ -1,1 +1,2 @@
-export * from "./WalletsCsv"
+export * as builders from "./builders"
+export * as sdls from "./sdls"
