+++ src/schema/sdls/ConsensusBountyHuntersCsv.ts
@@ -1,0 +1,7 @@
+export const ConsensusBountyHuntersCsv = `
+  type ConsensusBountyHuntersCsv implements Node {
+    username: String,
+    name: String,
+    score: Int
+  }
+`
