+++ src/schema/sdls/ExecutionBountyHuntersCsv.ts
@@ -1,0 +1,7 @@
+export const ExecutionBountyHuntersCsv = `
+  type ExecutionBountyHuntersCsv implements Node {
+    username: String,
+    name: String,
+    score: Int
+  }
+`
